rewrite to better support actual es6 import statements, add tests
It turns out that it's not as easy as I thought to convert from destructured imports to destructured variable assignments: the default export may be (and likely is!) an object with properties that have nothing to do with the named assignments.

So in our exported object, if necessary we include two different keys: the 'my-module' key contains the default export, and the 'my-module:obj' key contains any named exports.